# crud/docs/swagger.yaml
definitions:
  main.UserInfo:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /user:
    post:
      consumes:
      - application/json
      description: Create a new user with "username" and "password"
      parameters:
      - description: User info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.UserInfo'
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Invalid request body or empty username/password
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create a new user
      tags:
      - users
  /user/{id}:
    delete:
      description: Delete user with given "user_id"
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete user
      tags:
      - users
    get:
      description: Read user with given "user_id"
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User data with masked password
          schema:
            $ref: '#/definitions/main.UserInfo'
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Read user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user info with given userID, requires "username" and "password"
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: User info with 'username' and 'password'
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.UserInfo'
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Invalid request body or bad username/password
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Update user
      tags:
      - users
swagger: "2.0"